Phase 13: Cleanup & final polish
Compiler warning fixes:
- graph.hpp: Fixed sign-conversion warnings
  * Added explicit static_cast<std::size_t> for int→size_t conversions
  * Affects add_edge() and neighbors() methods

- bmssp.cpp: Marked debug counters as [[maybe_unused]]
  * relax_attempts, insert_count, k_count
  * Used for debugging but not always referenced

- bmssp_state.cpp: Marked width_bytes lambda as [[maybe_unused]]
  * Only used when ENABLE_BMSSP_VERIFIER is defined

.clang-tidy configuration:
- Updated with comprehensive checks for C++20
- Enabled: bugprone, cppcoreguidelines, modernize, performance, readability
- Disabled overly strict checks:
  * Magic numbers (algorithm has many paper constants)
  * Identifier length (short names like k, t, l, M from paper)
  * Function cognitive complexity (algorithms are inherently complex)
- Added naming convention rules:
  * lower_case for namespaces, functions, variables
  * CamelCase for classes, structs
  * Private member suffix: _
- Set WarningsAsErrors to '' (warnings only, not errors)

CHANGELOG.md:
- Comprehensive documentation of all 13 implementation phases
- Phase-by-phase breakdown:
  * Phase 0: Baseline & goals
  * Phase 1: Module scaffolding
  * Phase 2/2b: Distance representation & widening (32→64→128→BigInt)
  * Phase 3: FindPivots (Algorithm 1)
  * Phase 4: Data structure 𝒟 (Lemma 3.3)
  * Phase 5: BaseCase (Algorithm 2)
  * Phase 6: Recursive BMSSP (Algorithm 3)
  * Phase 7: Top-level driver with auto-parameters
  * Phase 8: Invariants & verification
  * Phase 9: Testing suite (43 tests)
  * Phase 10: Performance instrumentation
  * Phase 11: Constant-degree validation
  * Phase 12: Documentation
  * Phase 13: Cleanup

Technical highlights section:
- Distance widening design rationale
- Complexity achievement (O(m + n log^{2/3} n))
- Build configuration reference
- Testing summary

All 41/41 tests passing (100%)

+++ include/cpp_starter/bmssp/graph.hpp
@@ -18,10 +18,10 @@ class Graph {
     return adj_.size();
   }
   void add_edge(int u, int v, uint64_t w) {
-    adj_[u].push_back({v, w});
+    adj_[static_cast<std::size_t>(u)].push_back({v, w});
   }
   const std::vector<Edge>& neighbors(int u) const {
-    return adj_[u];
+    return adj_[static_cast<std::size_t>(u)];
   }
 
  private:
